+++ src/minecraft/dark/assembly/common/machine/BlockTurntable.java
@@ -63,22 +63,22 @@ public class BlockTurntable extends BlockAssembly
 
     public static int determineOrientation(World world, int x, int y, int z, EntityPlayer entityPlayer)
     {
-        if (MathHelper.abs((float) entityPlayer.posX - (float) x) < 2.0F && MathHelper.abs((float) entityPlayer.posZ - (float) z) < 2.0F)
+        if (MathHelper.abs((float) entityPlayer.posX - x) < 2.0F && MathHelper.abs((float) entityPlayer.posZ - z) < 2.0F)
         {
-            double var5 = entityPlayer.posY + 1.82D - (double) entityPlayer.yOffset;
+            double var5 = entityPlayer.posY + 1.82D - entityPlayer.yOffset;
 
-            if (var5 - (double) y > 2.0D)
+            if (var5 - y > 2.0D)
             {
                 return 1;
             }
 
-            if ((double) y - var5 > 0.0D)
+            if (y - var5 > 0.0D)
             {
                 return 0;
             }
         }
 
-        int var7 = MathHelper.floor_double((double) (entityPlayer.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+        int var7 = MathHelper.floor_double((entityPlayer.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
         return var7 == 0 ? 2 : (var7 == 1 ? 5 : (var7 == 2 ? 3 : (var7 == 3 ? 4 : 0)));
     }
 
@@ -138,7 +138,7 @@ public class BlockTurntable extends BlockAssembly
                     rotatable.setDirection(ForgeDirection.getOrientation(newDir));
 
                     world.markBlockForUpdate(position.intX(), position.intY(), position.intZ());
-                    world.playSoundEffect((double) x + 0.5D, (double) y + 0.5D, (double) z + 0.5D, "tile.piston.in", 0.5F, world.rand.nextFloat() * 0.15F + 0.6F);
+                    world.playSoundEffect(x + 0.5D, y + 0.5D, z + 0.5D, "tile.piston.in", 0.5F, world.rand.nextFloat() * 0.15F + 0.6F);
                 }
             }
             catch (Exception e)
